move targeting to where it originates, and make player tooltip more responsive
